requirements.txt
fastapi
uvicorn
pydantic
chromadb
groq
python-dotenv
requests
streamlit
pillow
beautifulsoup4
playwright
markdownify
pypdf
openai
sentence-transformers
